pyproject.toml
[project]
name = "prediction-market"
version = "0.1.0"
description = "Polymarket political market surveillance system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "websockets>=14.0",
    "aiosqlite>=0.21",
    "pandas>=2.2",
    "scipy>=1.14",
    "click>=8.1",
    "pydantic>=2.10",
    "tomli>=2.0; python_version < '3.11'",
    "tomli-w>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "respx>=0.22",
    "ruff>=0.9",
]

[project.scripts]
prediction-market = "prediction_market.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prediction_market"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
]

[tool.ruff]
target-version = "py312"
line-length = 100
